Refactoring and little changes to project structure before merging
