+++ src/main/java/codeemoji/inlay/external/scanners/OSSVulnerabilityScanner.java
@@ -1,5 +1,7 @@
-package codeemoji.inlay.external;
+package codeemoji.inlay.external.scanners;
 
+import codeemoji.inlay.external.DependencyInfo;
+import codeemoji.inlay.external.VulnerabilityInfo;
 import org.json.JSONArray;
 import org.json.JSONObject;
