+++ src/main/java/codeemoji/inlay/external/services/BaseVulnerabilityExternalService.java
@@ -1,7 +1,10 @@
-package codeemoji.inlay.external;
+package codeemoji.inlay.external.services;
 
 import codeemoji.core.external.CEExternalService;
 import codeemoji.core.util.CEUtils;
+import codeemoji.inlay.external.DependencyInfo;
+import codeemoji.inlay.external.VulnerabilityInfo;
+import codeemoji.inlay.external.scanners.VulnerabilityScanner;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.roots.libraries.Library;
 import com.intellij.openapi.roots.libraries.LibraryTable;
@@ -13,7 +16,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
-public abstract class BaseExternalServiceForVulnerability implements CEExternalService<VirtualFile, Object> {
+public abstract class BaseVulnerabilityExternalService implements CEExternalService<VirtualFile, Object> {
     protected final Map<VirtualFile, Object> persistedData = new HashMap<>();
     protected Map<DependencyInfo, List<VulnerabilityInfo>> vulnerabilityMap = new HashMap<>();
     protected DependencyInfo[] dependencyInfos;
