+++ src/main/java/codeemoji/inlay/structuralanalysis/element/method/ExternalFunctionalityInvokingMethod.java
@@ -4,16 +4,11 @@ import codeemoji.core.collector.simple.CEMethodCollector;
 import codeemoji.core.collector.simple.CEReferenceMethodCollector;
 import codeemoji.core.provider.CEProviderMulti;
 import codeemoji.core.util.CEUtils;
-import codeemoji.inlay.vulnerabilities.MethodAnalysisUtils;
 import com.intellij.codeInsight.hints.ImmediateConfigurable;
 import com.intellij.codeInsight.hints.InlayHintsCollector;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.roots.ProjectFileIndex;
-import com.intellij.psi.PsiJavaFile;
 import com.intellij.psi.PsiMethod;
-import com.intellij.psi.PsiMethodCallExpression;
-import com.intellij.psi.util.PsiTreeUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -65,7 +60,7 @@ public class ExternalFunctionalityInvokingMethod extends CEProviderMulti<Externa
 
         if(
                 externalFunctionalityInvokingMethods.length > 0 &&
-                        Arrays.stream(externalFunctionalityInvokingMethods).anyMatch(externalFunctionalityInvokingMethod -> MethodAnalysisUtils.checkMethodExternality(externalFunctionalityInvokingMethod, project))
+                        Arrays.stream(externalFunctionalityInvokingMethods).anyMatch(externalFunctionalityInvokingMethod -> CEUtils.checkMethodExternality(externalFunctionalityInvokingMethod, project))
         ){
             return true;
         }
