--- src/main/java/codeemoji/inlay/vulnerabilities/MethodAnalysisUtils.java
@@ -1,13 +1,0 @@
-package codeemoji.inlay.vulnerabilities;
-
-import codeemoji.core.util.CEUtils;
-import com.intellij.openapi.project.Project;
-import com.intellij.openapi.roots.ProjectFileIndex;
-import com.intellij.psi.PsiJavaFile;
-import com.intellij.psi.PsiMethod;
-import com.intellij.psi.PsiMethodCallExpression;
-import com.intellij.psi.util.PsiTreeUtil;
-
-public class MethodAnalysisUtils {
-
-}
